feat(router): implement routing
